+++ Packages/com.unity.cloud.gltfast/CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved comma placement in export summary (thanks anonymous for [#46](https://github.com/Unity-Technologies/com.unity.cloud.gltfast/pull/46)).
 - [Materials variants](xref:GLTFast.MaterialsVariantsComponent)' inspector shows correct variant when regaining focus (thanks [anonymous2585](https://github.com/anonymous2585) for [#48](https://github.com/Unity-Technologies/com.unity.cloud.gltfast/pull/48)).
 - (Import) Fixed merging of mesh primitives with shared vertex buffer, but different indices type.
+- (Import) Fixed import of meshes where only some primitives are Draco compressed. Compressed and uncompressed primitives with matching attributes were clustered together and the second kind was passed to the wrong mesh generator.
 
 ### Removed